+++ modules/forum/src/main/ui/ForumBits.scala
@@ -37,7 +37,7 @@ final class ForumBits(helpers: Helpers):
   val dataUnsub = attr("data-unsub")
 
   def postTextarea(field: play.api.data.Field)(modifiers: Modifier*)(using Translate, Context) =
-    bits.markdownEditor(lila.core.data.MarkdownRealm.forum):
+    bits.markdownEditor(MarkdownRealm.forum):
       form3.textarea(field, "post-text-area")(
         rows := 10,
         autocomplete := "off",
