+++ modules/forum/src/main/ui/PostUi.scala
@@ -132,7 +132,7 @@ final class PostUi(helpers: Helpers, bits: ForumBits):
         ctx.me.soUse[Option[Tag]]:
           post.shouldShowEditForm.option:
             postForm(cls := "edit-post-form none", action := routes.ForumPost.edit(post.id))(
-              lila.ui.bits.markdownEditor(lila.core.data.MarkdownRealm.forum):
+              lila.ui.bits.markdownEditor(MarkdownRealm.forum):
                 textarea(
                   bits.dataTopic := topic.id,
                   name := "changes",
