+++ modules/ui/src/main/router/LilaRouter.scala
@@ -5,10 +5,10 @@ import play.api.mvc.{ PathBindable, QueryStringBindable }
 import scalalib.newtypes.SameRuntime
 
 import lila.core.id.*
-import lila.core.data.MarkdownRealm
 import lila.core.study.StudyOrder as StudyOrder
 import lila.core.ublog.{ BlogsBy, QualityFilter as BlogQualityFilter }
 import lila.core.misc.AppealTopic
+import lila.ui.MarkdownRealm
 
 object LilaRouter:
